Changed/standardized some syntax throughout most of the code.
Most of the code is more wordy, but also more specific, now. Density now takes the index of the particle as an argument rather than its position.
Similar changes were made for density_comp and  density_arr. The only part of the code that retained its brevity is the leapfrog function, because
its just too pretty looking to tamper with.