Add bulk viscosity and other changes.
This is a running version of the code which seems to work correctly from
preliminary tests.

Here is a summary of changes:
  * Add bulk viscosity to the analysis
  * This includes adding the function Pi and modifying the energy
evolution functions.
  * Changes the way energy evolution is handled. The energy rates are now
passed on to the leapfrog integrator and dealt with there.
  * Leapfrog integration now handles energy evolution as well. See Cossins
equations 3.152 through 3.155.
  * Alters the indexing order of the main arrays so that time is now in
the first (slowest varying) axis.
  * Alters the calling signature "var_smoothlength_sim" so that only the
initial conditions are given as inputs (rather than full arrays with a
time axis).