Energy function defined
Finished setting up the energy function as well as the component functions for energy (internal energy, kinetic energy, gravitational potential), renamed the functions in the gravity module as they were named gravitational potential when they were the gravitational smoothing kernal functions and derivatives. Next steps are to write scripts that will test the conservation of energy and find instances where pi is negative.