Minor testing and Total Energy implementation
I found that the method of getting the initial velocities gave velocities that were too large and caused an overall expansion of the gas cloud. I've set the initial energies and velocities to 0 to fix that. However, there is now no rotations. I will further investigate this while trying to implement the total Energy function.

A problem I found with the total energy function is that the position array does not work in the grav_kernal function. I think this is just a implementation error that is easy to fix.