Energy delta testing completed
Generated two plots using test_script.py and saved them as jpg. Unfortunately some information is lost and to see the information the script must be run again. I tried to figure out a way to regenerate the scripts but due to the way that total energy is calculated and stored, and plotted, there is no easy way to do so. I will try to figure out a way but my focus will be on the next steps.

Removed the saved gifs, they weren't useful in this kind of testing.

finished creating test_script.py, this runs the simulation over the same time period but varying number of time steps. Then plots the change in energy over time and the energy delta of each run. I tried to save array values but it didn't quite work as expected.