Plots are now able to be generated from saved data
I figured out how to save the data, I was going for an overly complicated solution but the simpler solution is just as elegant. If the jpg's of the data are not enough you can use the plot_regenerator.py program to regenerate the plot for manipulation of the plots.